pip_requirements.txt
django
djangorestframework
pip
